Initial version of overridable KeyMaker
